+++ lib/elixir/lib/record.ex
@@ -29,9 +29,11 @@ defmodule Record do
     quote do
       defmodule unquote(name) do
         @moduledoc false
-        @type t :: tuple
         Record.deffunctions(unquote(values), unquote(opts), __ENV__)
         unquote(block)
+        unless Kernel.Typespec.defines_type?(__MODULE__, :t, 0) do
+          Kernel.Typespec.deftype t :: tuple
+        end
       end
     end
   end